+++ Source/Database/Database.swift
@@ -30,6 +30,7 @@ public class Database {
     ///
     /// - Parameters:
     ///   - storageLocation:             The storage location path to use during initialization.
+    ///   - tableLockPolicy:             The table lock policy used to handle table lock errors. `.fastFail` by default.
     ///   - multiThreaded:               Whether the database should be multi-threaded. `true` by default.
     ///   - sharedCache:                 Whether the database should use a shared cache. `false` by default.
     ///   - drainDelay:                  Total time to wait before draining available reader connections. `1.0` by 
@@ -41,6 +42,7 @@ public class Database {
     /// - Throws: A `SQLiteError` if SQLite encounters an error opening the writable connection.
     public init(
         storageLocation: StorageLocation = .inMemory,
+        tableLockPolicy: TableLockPolicy = .fastFail,
         multiThreaded: Bool = true,
         sharedCache: Bool = false,
         drainDelay: TimeInterval = 1.0,
@@ -50,6 +52,7 @@ public class Database {
     {
         let writerConnection = try Connection(
             storageLocation: storageLocation,
+            tableLockPolicy: tableLockPolicy,
             readOnly: false,
             multiThreaded: multiThreaded,
             sharedCache: sharedCache
@@ -61,6 +64,7 @@ public class Database {
 
         readerConnectionPool = ConnectionPool(
             storageLocation: storageLocation,
+            tableLockPolicy: tableLockPolicy,
             availableConnectionDrainDelay: drainDelay,
             connectionPreparation: readerConnectionPreparation
         )
@@ -74,6 +78,7 @@ public class Database {
     ///
     /// - Parameters:
     ///   - storageLocation:             The storage location path to use during initialization.
+    ///   - tableLockPolicy:             The table lock policy used to handle table lock errors. `.fastFail` by default.
     ///   - flags:                       The bitmask flags to use when initializing the database.
     ///   - drainDelay:                  Total time to wait before draining available reader connections. `1.0` by 
     ///                                  default.
@@ -84,19 +89,26 @@ public class Database {
     /// - Throws: A `SQLiteError` if SQLite encounters an error opening the writable connection.
     public init(
         storageLocation: StorageLocation,
+        tableLockPolicy: TableLockPolicy = .fastFail,
         flags: Int32,
         drainDelay: TimeInterval = 1.0,
         writerConnectionPreparation: ((Connection) throws -> Void)? = nil,
         readerConnectionPreparation: ((Connection) throws -> Void)? = nil)
         throws
     {
-        let writerConnection = try Connection(storageLocation: storageLocation, flags: flags)
+        let writerConnection = try Connection(
+            storageLocation: storageLocation,
+            tableLockPolicy: tableLockPolicy,
+            flags: flags
+        )
+
         try writerConnectionPreparation?(writerConnection)
 
         writerConnectionQueue = ConnectionQueue(connection: writerConnection)
 
         readerConnectionPool = ConnectionPool(
             storageLocation: storageLocation,
+            tableLockPolicy: tableLockPolicy,
             availableConnectionDrainDelay: drainDelay,
             connectionPreparation: readerConnectionPreparation
         )